+++ src/main/scala/org/dia/algorithms/mcc/NetcdfDFSMCC.scala
@@ -137,7 +137,7 @@ object NetcdfDFSMCC {
 
         def updateComponent(label: Double, frame: String, value: Double, row: Int, col: Int): Unit = {
           if (label != 0.0) {
-            val node = nodeMap.getOrElse(frame + ":" + label, new MCCNode(frame.toInt, label))
+            val node = nodeMap.getOrElse(frame + ":" + label, new MCCNode(frame, label))
             node.update(value, row, col)
             nodeMap.update((frame + ":" + label), node)
           }
